feat(tasks): Adapt `add-task`
- Rename task form to add task
- Use drawer to adapt on mobile screens
